requirements.txt
tensorboard
scikit-learn
seqeval
psutil
sacrebleu
rouge-score
tensorflow_datasets
pytorch-lightning==0.8.5
matplotlib
git-python==1.0.3
faiss
streamlit
elasticsearch
pandas
nlp
fire
pytest
conllu
tokenizers==0.9.0